+++ Pokemon/main.cpp
@@ -1,15 +1,13 @@
 #include<iostream>
 #include<string>
+#include <limits>
+//
 #include "PokemonType.hpp"
 #include "PokemonChoice.hpp"
 #include "Utility.hpp"
-#include <limits>
-#include <string>
+#include "Player.hpp"
+
 using namespace std;
-
-
-
-
 
 class Pokemon {
 public:
@@ -46,47 +44,6 @@ public:
         Utility::waitForEnter();
     }
 
-};
-
-class Player {
-public:
-    std::string name;
-    Pokemon chosenPokemon;
-
-    Player() {
-        name = "Trainer";
-        chosenPokemon = Pokemon(/*"Pikachu", PokemonType::ELECTRIC, 10*/);
-    }
-
-    Player(std::string p_name, const Pokemon& p_chosenPokemon) {
-        name = p_name;
-        chosenPokemon = p_chosenPokemon;
-    }
-
-    Player(const Player& player) {
-        name = player.name;
-        chosenPokemon = player.chosenPokemon;
-    }
-
-    void choosePokemon(int choice) {
-        switch ((PokemonChoice)choice)
-        {
-        case PokemonChoice::CHARMANDER:
-            chosenPokemon = Pokemon("CHARMANDER", PokemonType::FIRE, 100);
-            break;
-        case PokemonChoice::BULBASAUR:
-            chosenPokemon = Pokemon("BULBASAUR", PokemonType::GRASS, 100);
-            break;
-        case PokemonChoice::SQUIRTLE:
-            chosenPokemon = Pokemon("SQUIRTLE", PokemonType::WATER, 100);
-            break;
-        default:
-            chosenPokemon = Pokemon("PIKACHU", PokemonType::ELECTRIC, 100);
-            break;
-        }
-        cout << "Player: " << name << " chose " << chosenPokemon.name << "!\n";
-        Utility::waitForEnter();
-    }
 };
 
 class ProfessorOak {
@@ -156,7 +113,7 @@ public:
     }
 };
 
-void gameLoop(Player player) {
+void gameLoop(Player &player) {
     int choice;
     bool keepPlaying = true;
     while (keepPlaying) {
